Adding support for pictures
1. Back Image: Here you can provide a url to an image and it will use it
as the background instead of the usual gradient and background icon.

2. Images in Content: Here you have a new verb "picture" which allows
you to add an image to the content of your card and scale it to fit
appropriately. The syntax is "picture | url | heightInPixels"

3. More Colors: Because you can now use any image as a background, I ran
into the problem of the default image clashing terribly with the images
I was using. So I added all the other colors to the color picker and
increased the maxWidth of the picker to fit all the new colors. I could
have been more picky about which colors because there are a lot of
duplicates with different names, but I didn't have the time for it.

4. It looks like I made some changes to bootstrap.min.css. It was a
while back and I don't remember what it does. However I was able to use
a better diffing tool than the one which comes with github and was able
to tell that it's a hack of a change made to .dropdown-menu changing
"min-width:160px" to "width:640px"

It's been a while and I really can't recall what that's for; it's
probably for the colorpicker but if it is, I don't really know of a
clean way to do it. I'll test it to see if it's even necessary and
remove it if it isn't. Let me know your thoughts on this as I'm not
really a web developer so I don't know if it's ok to muck with this. My
general strategy for UI is to change things with the developer toolbar
and then make the changes to the corresponding file if it looks good, so
feel free to tell me if my code sucks, and if it should be improved in
any way.